Added release workflows
2 new workflows, 1 to create a release and 1 to publish the package to
NPM. The release workflow tests and builds the tarballs before creating
a draft release when a new tag matching the pattern v*.*.* is created.
This draft release can then be manually checked before being published.
The npm workflow then, when a release is published, builds the NPM
package from the release to NPM.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>