Validate organisation ID
Now validate the organisation ID when setting with config org set by
making a call to the project/organization/<id> endpoint. If this
endpoint returns anything except 2xx we say that the ID is invalid. To
ensure we don't mask other errors we also log the error thrown in the
debug log.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>