Validate project ID
Now validate the project ID when setting with config project set by
making a call to the project/ endpoint setting the correct header. If
this endpoint returns anything except 2xx we say that the ID is invalid.
To ensure we don't mask other errors we also log the error thrown in the
debug log.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>