Validate project and organisation IDs (#78)
* Validate organisation ID

Now validate the organisation ID when setting with config org set by
making a call to the project/organization/<id> endpoint. If this
endpoint returns anything except 2xx we say that the ID is invalid. To
ensure we don't mask other errors we also log the error thrown in the
debug log.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>

* Validate project ID

Now validate the project ID when setting with config project set by
making a call to the project/ endpoint setting the correct header. If
this endpoint returns anything except 2xx we say that the ID is invalid.
To ensure we don't mask other errors we also log the error thrown in the
debug log.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>

* Improved org and project detection

Instead of comparing the string of the error message, we now instead
have a custom error that includes the response status for more reliable
detection of invalid IDs

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>

---------

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>