Switched from node-fetch to http-call
In order to fix #107 and #87, the package used to make API calls has
been switched from node-fetch to http-call, a package developed by
Heroku for their CLI that supports automatic retries and proxy handling.

The API class has been fully updated to make use of the new package.
This did result in a slight change to the interface of the class with
the plain flag being moved into the options type and the body now can
be directly passed as an object instead of having to convert to a JSON
string first. This however has not had much impact on the rest of the
code and the minor change has been made where required.

Errors from API requests are now of the type HTTPError rather than
ApiError (our own custom error that has now been removed). This provides
more information about the error including the response body so the
parts of code that dealt previously with ApiError have been updated for
HTTPError and the old ApiError has been removed.

Whilst testing http-call for the Auth class, an unrelated bug was
discovered in the pollToken method. The sleep call passed the time to
sleep in seconds when milliseconds was expected, resulting in too many
requests to the token endpoint and the resulting slow_down requests
being ignored.

A further logic error with the pollToken method was discovered when
handling the slow_down error. The intention was for the sleep delay to
double each time the slow_down error was handled, however as implemented
the original retry interval was just doubled, i.e. 5, 10, 10, ..., 10.
this was fixed along side the millisecond issue.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>