Implemented #113
If the current project is set, we get it's details from the API to present
it's name to the user.
Note: this is technically a breaking change as we change the schema for
the --json output, as well as alter the name of the Current Project ID
row, but we are still in 0.x.x so it is to be expected.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>