Added delete project (#119)
The delete project command has been implemented to allow users to delete
a project from an organisation. If there are any roles or accounts with
access to the project, the user will receive an error stating  that
these must be removed first.

If the user does not specify the '-p' flag, then the current working
project will be deleted and set to undefined. The user will be informed
of the need for them to set a new current working project.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>