Allow user to specify user using email (#122)
* Allow user to specify user using email

Fix #94

Now for the commands where the user input an ID for operations, an email
is also supported. Emails can only be used if the user who's email it
is is a member of the organisation. If the user is not a member of the
organisation, this will not work and an error will be returned.

It is possible for multiple users to share the same email (e.g. social
login). In this case, if there are two accounts with the same email on
an org, the operation will also fail.

There has been a slight change to the flags of the
projects:accounts:edit command as a result of this change. Now there is
an optional organization-id flag that allows the user to set the
current org inline. This is because this command now makes use of an
organisation API.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>

* Get org ID from current project

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>

---------

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>