Switched to use OIDC .well-known config
In order to reduce the number of environment variables we use, now we
can set the auth server using the APIMETRICS_AUTH_SERVER env var instead
of having to set each endpoint individually. This is then used to access
the .well-known/openid-configuration endpoint to get the required
endpoint info. The call to this endpoint is only made if the auth server
is different from the default in order to avoid making unnecessary API
calls.

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>