Add `@oclif/plugin-warn-if-update-available`
This will allow users to be alerted when a new version of the CLI is released. The CLI will poll the registry a maximum of once every 7 days when the user runs a command. If the user is running an outdated version, a warning will be printed to STDERR. It should be noted that this warning will be printed even if --json is passed so users should make sure to only read from STDOUT if they are fetching the data (this should be the default for any ordinary use case).

Signed-off-by: Matthew Nickson <matthew@apimetrics.com>